backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
ansi2html==1.9.2
anyio==4.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.3.0
billiard==4.2.1
Brotli==1.1.0
celery==5.3.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==44.0.0
cssselect2==0.7.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
Django==5.1.2
django-allauth==0.55.0
django-celery-beat==2.7.0
django-cors-headers==4.5.0
django-debug-toolbar==4.2.0
django-environ==0.10.0
django-filter==23.2
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.0
drf-yasg==1.21.5
exceptiongroup==1.2.2
Flask==2.2.5
fonttools==4.55.3
freetype-py==2.3.0
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
langchain==0.3.13
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langsmith==0.2.6
lxml==5.3.0
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
orjson==3.10.13
packaging==24.2
pillow==11.0.0
plotly==5.24.1
pluggy==1.5.0
prometheus-client==0.16.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg==3.2.3
pycairo==1.27.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydyf==0.11.0
PyJWT==2.10.1
PyPDF2==3.0.1
pyphen==0.17.0
pytest==7.4.0
pytest-django==4.5.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==0.8.11
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.0
referencing==0.35.1
reportlab==4.0.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rlPyCairo==0.3.0
rpds-py==0.22.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
sentry-sdk==1.32.0
setuptools==75.7.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
tenacity==9.0.0
tinycss2==1.4.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==57.1
webencodings==0.5.1
Werkzeug==2.2.3
whitenoise==6.5.0
XlsxWriter==3.1.0
yarl==1.18.3
zopfli==0.2.3.post1
